# implementation/virt-reader/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "AJIN RFID Tracking API",
        "description": "RFID 기반 물류 추적 시스템 API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/dashboard/summary": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Get Dashboard Summary",
                "description": "대시보드 요약 정보 (권한: dashboard:read)\n\n- 활성 팔레트 수\n- 총 재고 수량\n- 금일 생산량\n- 리더기 연결 상태",
                "operationId": "get_dashboard_summary_api_v1_dashboard_summary_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DashboardSummary"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ]
            }
        },
        "/api/v1/dashboard/process-status": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Get Process Status",
                "description": "공정별 현황 (권한: dashboard:read)\n\n각 공정의 활성 팔레트 수와 상태별 분포를 반환합니다.\n(N+1 쿼리 최적화: 단일 쿼리로 집계)",
                "operationId": "get_process_status_api_v1_dashboard_process_status_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProcessStatusList"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ]
            }
        },
        "/api/v1/dashboard/readers": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Get Reader Status",
                "description": "리더기 상태 조회 (권한: dashboard:read)\n\n모든 등록된 리더기의 연결 상태와 마지막 스캔 시간을 반환합니다.\nlast_scan_time은 rfid_reader_locations 테이블에 직접 저장되어 조인 없이 빠르게 조회됩니다.",
                "operationId": "get_reader_status_api_v1_dashboard_readers_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ReaderStatusList"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ]
            }
        },
        "/api/v1/dashboard/inventory/stock": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Get Stock Inventory",
                "description": "재고 현황 (FIFO 관리용) (권한: dashboard:read)\n\n품번/공정별 재고를 생산일자순으로 정렬하여 반환합니다.\n경과일에 따른 긴급도 표시 포함.",
                "operationId": "get_stock_inventory_api_v1_dashboard_inventory_stock_get",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "part_number",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Part Number"
                        }
                    },
                    {
                        "name": "process_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "integer"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Process Id"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "description": "정렬 기준",
                            "default": "production_date",
                            "title": "Sort"
                        },
                        "description": "정렬 기준"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StockInventoryResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/dashboard/recent-activities": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Get Recent Activities",
                "description": "최근 활동 이력 (권한: dashboard:read)\n\n최근 팔레트 상태 변경 이력을 반환합니다.",
                "operationId": "get_recent_activities_api_v1_dashboard_recent_activities_get",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 100,
                            "minimum": 1,
                            "default": 20,
                            "title": "Limit"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "DashboardSummary": {
                "properties": {
                    "active_pallets": {
                        "type": "integer",
                        "title": "Active Pallets",
                        "description": "활성 팔레트 수"
                    },
                    "total_stock": {
                        "type": "integer",
                        "title": "Total Stock",
                        "description": "총 재고 수량"
                    },
                    "today_production": {
                        "type": "integer",
                        "title": "Today Production",
                        "description": "금일 생산량"
                    },
                    "reader_status": {
                        "additionalProperties": {
                            "type": "integer"
                        },
                        "type": "object",
                        "title": "Reader Status",
                        "description": "리더기 상태 (connected, total)"
                    }
                },
                "type": "object",
                "required": [
                    "active_pallets",
                    "total_stock",
                    "today_production",
                    "reader_status"
                ],
                "title": "DashboardSummary",
                "description": "대시보드 요약 정보"
            },
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "LotStock": {
                "properties": {
                    "lot_no": {
                        "type": "string",
                        "title": "Lot No"
                    },
                    "pallet_no": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Pallet No"
                    },
                    "production_date": {
                        "type": "string",
                        "format": "date",
                        "title": "Production Date"
                    },
                    "days_old": {
                        "type": "integer",
                        "title": "Days Old"
                    },
                    "quantity": {
                        "type": "integer",
                        "title": "Quantity"
                    },
                    "status": {
                        "type": "string",
                        "title": "Status",
                        "description": "urgent, warning, normal"
                    }
                },
                "type": "object",
                "required": [
                    "lot_no",
                    "production_date",
                    "days_old",
                    "quantity",
                    "status"
                ],
                "title": "LotStock",
                "description": "LOT별 재고 정보"
            },
...
            "ProcessStatus": {
                "properties": {
                    "process_id": {
                        "type": "integer",
                        "title": "Process Id"
                    },
                    "process_name": {
                        "type": "string",
                        "title": "Process Name"
                    },
                    "production_line": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Production Line"
                    },
                    "active_pallets": {
                        "type": "integer",
                        "title": "Active Pallets",
                        "description": "활성 팔레트 수",
                        "default": 0
                    },
                    "status_breakdown": {
                        "additionalProperties": {
                            "type": "integer"
                        },
                        "type": "object",
                        "title": "Status Breakdown",
                        "description": "상태별 팔레트 수"
                    }
                },
                "type": "object",
                "required": [
                    "process_id",
                    "process_name"
                ],
                "title": "ProcessStatus",
                "description": "공정별 상태"
            },
            "ProcessStatusList": {
                "properties": {
                    "processes": {
                        "items": {
                            "$ref": "#/components/schemas/ProcessStatus"
                        },
                        "type": "array",
                        "title": "Processes"
                    },
                    "total_active_pallets": {
                        "type": "integer",
                        "title": "Total Active Pallets"
                    },
                    "last_updated": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Last Updated"
                    }
                },
                "type": "object",
                "required": [
                    "processes",
                    "total_active_pallets",
                    "last_updated"
                ],
                "title": "ProcessStatusList",
                "description": "공정별 현황 목록"
            },
...
            "ReaderStatus": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    },
                    "port_name": {
                        "type": "string",
                        "title": "Port Name"
                    },
                    "process_name": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Process Name"
                    },
                    "location_type": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Location Type"
                    },
                    "status": {
                        "type": "string",
                        "title": "Status",
                        "description": "CONNECTED, DISCONNECTED, ERROR"
                    },
                    "last_scan_time": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "date-time"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Last Scan Time"
                    },
                    "is_active": {
                        "type": "boolean",
                        "title": "Is Active",
                        "default": true
                    },
                    "error": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Error"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "port_name",
                    "status"
                ],
                "title": "ReaderStatus",
                "description": "리더기 상태"
            },
...
            "ReaderStatusList": {
                "properties": {
                    "readers": {
                        "items": {
                            "$ref": "#/components/schemas/ReaderStatus"
                        },
                        "type": "array",
                        "title": "Readers"
                    }
                },
                "type": "object",
                "required": [
                    "readers"
                ],
                "title": "ReaderStatusList",
                "description": "리더기 상태 목록"
            },
...
            "StockInventoryResponse": {
                "properties": {
                    "stock_items": {
                        "items": {
                            "$ref": "#/components/schemas/StockItem"
                        },
                        "type": "array",
                        "title": "Stock Items"
                    }
                },
                "type": "object",
                "required": [
                    "stock_items"
                ],
                "title": "StockInventoryResponse",
                "description": "재고 현황 응답"
            },
            "StockItem": {
                "properties": {
                    "item_code": {
                        "type": "string",
                        "title": "Item Code"
                    },
                    "item_name": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Item Name"
                    },
                    "vehicle_model": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Vehicle Model"
                    },
                    "process_name": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Process Name"
                    },
                    "production_line": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Production Line"
                    },
                    "lots": {
                        "items": {
                            "$ref": "#/components/schemas/LotStock"
                        },
                        "type": "array",
                        "title": "Lots",
                        "default": []
                    }
                },
                "type": "object",
                "required": [
                    "item_code"
                ],
                "title": "StockItem",
                "description": "품번별 재고 정보"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    },
                    "input": {
                        "title": "Input"
                    },
                    "ctx": {
                        "type": "object",
                        "title": "Context"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        },
        "securitySchemes": {
            "OAuth2PasswordBearer": {
                "type": "oauth2",
                "flows": {
                    "password": {
                        "scopes": {},
                        "tokenUrl": "/api/v1/auth/login"
                    }
                }
            }
        }
    }
}